+++ db/migrate/20220914235036_add_risograph_to_medium.rb
@@ -4,10 +4,14 @@
 
 class AddRisographToMedium < ActiveRecord::Migration[7.0]
   def up
+    return if Rails.env.test?
+
     Term.create!(value: 'Risograph', facet: facet_medium)
   end
 
   def down
+    return if Rails.env.test?
+
     term = Term.find_by!(value: 'Risograph', facet: facet_medium)
     term.destroy!
   end
